Make TreeTabSessionObserver
