+++ browser/ui/browser_window/internal/BUILD.gn
@@ -40,6 +40,7 @@ source_set("internal") {
     deps += [
       "//brave/browser/ui/focus_mode",
       "//brave/browser/ui/tabs:tab_menu",
+      "//brave/browser/ui/tabs:tree_tab_session_observer",
       "//brave/browser/ui/views/page_info",
       "//chrome/browser/ui/tabs:features",
     ]
