+++ browser/ui/tabs/tree_tab_session_observer.cc
@@ -1,0 +1,145 @@
+// Copyright (c) 2026 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
+#include "brave/browser/ui/tabs/tree_tab_session_observer.h"
+
+#include <string>
+
+#include "base/check.h"
+#include "base/feature_list.h"
+#include "brave/browser/sessions/brave_session_keys.h"
+#include "brave/components/tabs/public/tree_tab_node.h"
+#include "brave/components/tabs/public/tree_tab_node_tab_collection.h"
+#include "chrome/browser/sessions/session_service.h"
+#include "chrome/browser/sessions/session_service_factory.h"
+#include "chrome/browser/ui/tabs/features.h"
+#include "chrome/browser/ui/tabs/tab_strip_model.h"
+#include "components/sessions/content/session_tab_helper.h"
+#include "components/tabs/public/tab_interface.h"
+#include "content/public/browser/web_contents.h"
+
+namespace {
+
+// Finds the nearest TreeTabNodeTabCollection ancestor of |tab| and returns it,
+// or nullptr if the tab is not inside a tree node.
+const tabs::TreeTabNodeTabCollection* GetTreeTabNodeCollection(
+    const tabs::TabInterface* tab) {
+  const tabs::TabCollection* parent = tab->GetParentCollection();
+  if (parent && parent->type() == tabs::TabCollection::Type::TREE_NODE) {
+    return static_cast<const tabs::TreeTabNodeTabCollection*>(parent);
+  }
+  return nullptr;
+}
+
+}  // namespace
+
+TreeTabSessionObserver::TreeTabSessionObserver(Profile* profile,
+                                               TabStripModel* tab_strip_model,
+                                               SessionID session_id)
+    : profile_(profile),
+      tab_strip_model_(tab_strip_model),
+      session_id_(session_id) {
+  CHECK(profile_);
+  CHECK(tab_strip_model_);
+  tab_strip_model_->AddObserver(this);
+}
+
+TreeTabSessionObserver::~TreeTabSessionObserver() {
+  tab_strip_model_->RemoveObserver(this);
+}
+
+void TreeTabSessionObserver::OnTreeTabChanged(const TreeTabChange& change) {
+  CHECK(base::FeatureList::IsEnabled(tabs::kBraveTreeTab));
+
+  switch (change.type) {
+    case TreeTabChange::Type::kNodeCreated:
+      UpdateTreeTabSessionDataForNode(change.GetCreatedChange().node.get());
+      break;
+    case TreeTabChange::Type::kNodeCollapsedStateChanged:
+      UpdateTreeTabCollapsedState(
+          change.GetCollapsedStateChangedChange().node.get());
+      break;
+    case TreeTabChange::Type::kNodeReparented:
+      UpdateTreeTabSessionDataForNode(change.GetReparentedChange().node.get());
+      break;
+    case TreeTabChange::Type::kNodeWillBeDestroyed:
+      // Do not clear tree data when a node is destroyed. The tree node data
+      // must remain in the session so that "restore closed tab" can place the
+      // tab back into its original tree position. When the tab is actually
+      // closed, SessionService::TabClosed() removes the whole tab from the
+      // session. On browser restart a full rebuild will naturally omit tree
+      // data for any tabs that are no longer in tree nodes.
+      break;
+  }
+}
+
+void TreeTabSessionObserver::UpdateTreeTabSessionDataForNode(
+    const tabs::TreeTabNode& node) {
+  SessionService* const session_service =
+      SessionServiceFactory::GetForProfileIfExisting(profile_);
+  if (!session_service) {
+    return;
+  }
+
+  auto parent_id = node.GetParentTreeNodeId();
+  const std::string parent_id_str = parent_id ? parent_id->ToString() : "";
+  const std::string node_id_str = node.id().ToString();
+
+  for (const tabs::TabInterface* tab_iface : node.GetTabs()) {
+    auto* wc = tab_iface->GetContents();
+    CHECK(wc);
+    auto* session_helper = sessions::SessionTabHelper::FromWebContents(wc);
+    CHECK(session_helper);
+
+    const SessionID tab_id = session_helper->session_id();
+    const auto* tree_coll = GetTreeTabNodeCollection(tab_iface);
+    if (!tree_coll) {
+      // TODO(https://github.com/brave/brave-browser/issues/49792): Add support
+      // Tabs are in the groups or splits. In this case, we don't do anything
+      // for now.
+      continue;
+    }
+
+    CHECK_EQ(tree_coll->current_value_type(),
+             tabs::TreeTabNodeTabCollection::CurrentValueType::kTab);
+    session_service->AddTabExtraData(session_id_, tab_id, kBraveTreeNodeIdKey,
+                                     node_id_str);
+    session_service->AddTabExtraData(session_id_, tab_id,
+                                     kBraveTreeParentNodeIdKey, parent_id_str);
+    session_service->AddTabExtraData(session_id_, tab_id,
+                                     kBraveTreeNodeCollapsedKey,
+                                     node.collapsed() ? "1" : "0");
+  }
+}
+
+void TreeTabSessionObserver::UpdateTreeTabCollapsedState(
+    const tabs::TreeTabNode& node) {
+  SessionService* const session_service =
+      SessionServiceFactory::GetForProfileIfExisting(profile_);
+  if (!session_service) {
+    return;
+  }
+
+  // Only the primary (kTab) tab of a node carries the collapsed state key.
+  for (const tabs::TabInterface* tab_iface : node.GetTabs()) {
+    const auto* tree_coll = GetTreeTabNodeCollection(tab_iface);
+    if (!tree_coll) {
+      // TODO(https://github.com/brave/brave-browser/issues/49792): Add support
+      // Tabs are in the groups or splits. In this case, we don't do anything
+      // for now.
+      continue;
+    }
+
+    auto* wc = tab_iface->GetContents();
+    CHECK(wc);
+
+    auto* session_helper = sessions::SessionTabHelper::FromWebContents(wc);
+    CHECK(session_helper);
+
+    session_service->AddTabExtraData(session_id_, session_helper->session_id(),
+                                     kBraveTreeNodeCollapsedKey,
+                                     node.collapsed() ? "1" : "0");
+  }
+}
